Move personal details to a separate component
